Rename DeveloperTab to ApiSandbox and refactor component structure
