+++ src/components/ApiSandbox/ResponseDetails.tsx
